controller/requirements.txt
boto3==1.34.19
python-dotenv==1.0.1
requests==2.31.0
tensorflow==2.15.0
setuptools==69.0.2
wheel==0.42.0
urllib3==1.26.18
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7